join

+++ app/Http/Controllers/ArticleController.php
@@ -2,30 +2,75 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Resources\Article as ArticleResourse;
+use App\Http\Resources\ArticleCollection;
 use App\Article;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 class ArticleController extends Controller
 {
+    //private static $rules = [
+        //'title' => 'required|string|unique:articles|max:255',
+        //'body' => 'required',
+        //'category_id' => 'required|exists:categories',
+    //];
+    private static $messages = [
+        'required' => 'El campo :attribute es obligatorio.',
+        'body.required' => 'El body no es valido.',
+
+    ];
     public function index()
     {
-        return Article::all();
+        //$this->authorize('viewAny', Article::class);
+        return new ArticleCollection(Article::paginate(25));
     }
     public function show(Article $article)
     {
-        return $article;
+        $this->authorize('view', $article);
+        return response()->json(new ArticleResource($article), 200);
+    }
+    public function image(Article $article)
+    {
+        return response()->download(public_path(Storage::url($article->image)),
+            $article->title);
     }
     public function store(Request $request)
     {
-        return response()->json($article, 201);
+        $this->authorize('create', Article::class);
+
+        $request->validate([
+            'title' => 'required|string|unique:articles|max:255',
+            'body' => 'required',
+            'category_id' => 'required|exists:categories',
+            'image' => 'required|image|dimensions:min_width=200,min_height=200',
+        ], self::$messages);
+        $article = new Article($request->all());
+        $path = $request->image->store('public/articles');
+    //    $path = $request->image->store('public/articles', $request->user()->id . '_' . $article->title . '.' . $request->image->extension()););
+
+        $article->image= $path;
+        $article->save();
+        //$article = Article::create($request->all());
+        return response()->json(new ArticleResourse($article), 201);
     }
     public function update(Request $request, Article $article)
     {
+        $this->authorize('update', $article);
+
+        $request->validate([
+            'title' => 'required|string|unique:articles,title'.$article->id.'|max:255',
+            'body' => 'required',
+            'category_id' => 'required|exists:categories',
+        ],self::$messages);
+
         $article->update($request->all());
         return response()->json($article, 200);
     }
     public function delete(Request $request, Article $article)
     {
+        $this->authorize('delete', $article);
+
         $article->delete();
         return response()->json(null, 204);
     }
